Move code to read autoFilter to sheet part reader
